requirements.txt
python-dotenv
requests
pandas
pathlib
fastapi
uvicorn
joblib
scikit-learn==1.6.1